fix(profile): code review指摘修正 - handleFollowToggle・型重複・定数化
- handleFollowToggle のダミー遅延を空の async 関数に変更
- PublicProfile 型を packages/types に移動し両アプリから共有インポート
- paddingBottom マジックナンバーを SCROLL_BOTTOM_PADDING 定数に抽出
