+++ packages/types/src/index.ts
@@ -4,4 +4,4 @@ export type { Notification } from "./notification";
 export type { Summary } from "./summary";
 export type { Tag } from "./tag";
 export type { Translation } from "./translation";
-export type { User } from "./user";
+export type { PublicProfile, User } from "./user";
